added token verification2
