+++ src/main/kotlin/no/nav/tilleggsstonader/sak/behandling/BehandlingController.kt
@@ -87,30 +87,6 @@ class BehandlingController(
             identStønadstype.stønadstype,
         )
     }
-    @PostMapping("harBehandling")
-    @ProtectedWithClaims(issuer = "azuread", claimMap = ["roles=access_as_application"])
-    fun hentBehandlingStatusForPersonMedStønadstype(@RequestBody identStønadstype: IdentStønadstype): Boolean {
-        tilgangService.validerTilgangTilPersonMedBarn(identStønadstype.ident, AuditLoggerEvent.ACCESS)
-        logger.info("hello the fødselsnummmer is fødselsnummer" +identStønadstype.ident,)
-        logger.info("hello the fødselsnummmer is fødselsnummer" +identStønadstype.stønadstype,)
-        val behandlinger = fagsakService.hentBehandlingerForPersonOgStønadstype(
-            identStønadstype.ident,
-            identStønadstype.stønadstype,
-        )
-        logger.info("behandlinger is" +behandlinger[0])
-        if (behandlinger.isNotEmpty()) {
-            val validStatuses = listOf(
-                BehandlingStatus.OPPRETTET,
-                BehandlingStatus.UTREDES,
-                BehandlingStatus.FATTER_VEDTAK,
-                BehandlingStatus.SATT_PÅ_VENT,
-            )
-            val behandlingStatus = behandlinger.map { it.status }
-            logger.info("behandlinger status" +behandlingStatus)
-            return behandlingStatus.any { it in validStatuses }
-        }
-        return false
-    }
 
 
     @PostMapping("{behandlingId}/henlegg")
